refactor: renamed readme example element
