--- src/components/feature/example/README.md
@@ -1,48 +1,0 @@
-# class: `Example`
-
-## Mixins
-
-| Name             | Module | Package                    |
-| ---------------- | ------ | -------------------------- |
-| `ColorSchemable` |        | @/utils/color-scheme.utils |
-
-## Methods
-
-| Name     | Privacy   | Description | Parameters | Return           | Inherited From |
-| -------- | --------- | ----------- | ---------- | ---------------- | -------------- |
-| `render` | protected |             |            | `TemplateResult` |                |
-
-## CSS Properties
-
-| Name                               | Default | Description                               |
-| ---------------------------------- | ------- | ----------------------------------------- |
-| `--wcp-example-spacing`            |         | Inner padding of the example              |
-| `--wcp-example-border-radius`      |         | Border radius of the example              |
-| `--wcp-example-border-width`       |         | Border width of the example               |
-| `--wcp-example-dark-border-color`  |         | Border color of the example in dark mode  |
-| `--wcp-example-light-border-color` |         | Border color of the example in light mode |
-
-## Slots
-
-| Name      | Description              |
-| --------- | ------------------------ |
-| `code`    | Code example             |
-| `preview` | Rendered example preview |
-
-<hr/>
-
-# Variables
-
-| Name           | Description | Type     |
-| -------------- | ----------- | -------- |
-| `EXAMPLE_TABS` |             | `object` |
-
-<hr/>
-
-# Functions
-
-| Name | Description | Parameters | Return |
-| ---- | ----------- | ---------- | ------ |
-|      |             |            |        |
-
-<hr/>
