--- src/components/feature/example/example.component.scss
@@ -1,27 +1,0 @@
-@use '@/styles/utils' as utils;
-
-// prettier-ignore
-:host {
-  ---wcp-example-spacing: var(--wcp-example-spacing, #{0 utils.size(1)});
-  ---wcp-example-border-radius: var(--wcp-example-border-radius, var(--wcp-border-radius));
-  ---wcp-example-border-width: var(--wcp-example-border-width, 1px);
-
-  ---wcp-example-dark-border-color: var(--wcp-example-dark-border-color, var(--wcp-color-grey-24));
-  ---wcp-example-light-border-color: var(--wcp-example-light-border-color, var(--wcp-color-grey-80));
-
-  @include utils.color-scheme() using ($color-scheme) {
-    ---wcp-example-border-color: var(---wcp-example-#{$color-scheme}-border-color);
-  }
-}
-
-:host {
-  display: block;
-  padding: var(---wcp-example-spacing);
-
-  border: var(---wcp-example-border-width) solid var(---wcp-example-border-color);
-  border-radius: var(---wcp-example-border-radius);
-}
-
-slot[name='code']::slotted(pre) {
-  margin: 0;
-}
