+++ src/components/feature/readme-example/README.md
@@ -1,0 +1,23 @@
+# wcp-readme-example
+
+Shows an inline code example and a preview of the element in the readme.
+This is used in the markdown formatter to render `html` examples.
+
+**Mixins:** ColorSchemable
+
+## Slots
+
+| Name      | Description              |
+|-----------|--------------------------|
+| `code`    | Code example             |
+| `preview` | Rendered example preview |
+
+## CSS Custom Properties
+
+| Property                                  | Description                               |
+|-------------------------------------------|-------------------------------------------|
+| `--wcp-readme-example-border-radius`      | Border radius of the example              |
+| `--wcp-readme-example-border-width`       | Border width of the example               |
+| `--wcp-readme-example-dark-border-color`  | Border color of the example in dark mode  |
+| `--wcp-readme-example-light-border-color` | Border color of the example in light mode |
+| `--wcp-readme-example-spacing`            | Inner padding of the example              |
