+++ src/index.ts
@@ -2,13 +2,13 @@
  * @file Automatically generated by barrelsby.
  */
 
-export * from './components/feature/example/example.component';
 export * from './components/feature/navigation/navigation.component';
 export * from './components/feature/navigation-item/navigation-item.component';
 export * from './components/feature/preview-controls/preview-controls.component';
 export * from './components/feature/preview-details/preview-details.component';
 export * from './components/feature/preview-frame/preview-frame.component';
 export * from './components/feature/readme/readme.component';
+export * from './components/feature/readme-example/readme-example.component';
 export * from './components/feature/toggle-color-scheme/toggle-color-scheme.component';
 export * from './components/feature/toggle-sidebar/toggle-sidebar.component';
 export * from './components/layout/aside/aside.component';
